Fold scopes uniformly for repeat, reveal, and choice
This factors code out of repeat.js and into scope.js such that it can be
used by repeat, reveal, and choice. All of these primitives now destroy
scopes when they are removed from the document instead of attempting to
reuse them. The intention is to introduce a <reuse> tag when it becomes
necessary.

- repeat introduces iteration scopes.
- reveal introduces revelation scopes.
- choose introduces choice schopes by the name of the choice.

All introduced scopes cut through transitive caller scopes with the same
export renaming semantics.